+++ docs/src/@primer/gatsby-theme-doctocat/nav.yml
@@ -44,8 +44,6 @@
 - title: Components
   url: /components
   children:
-    - title: Action List
-      url: /components/action-list
     - title: Alerts
       url: /components/alerts
     - title: Autocomplete
