+++ docs/src/stories/components/ActionList/ActionListItemFeatures.stories.jsx
@@ -1,11 +1,37 @@
 import React from 'react'
 import clsx from 'clsx'
-import {ListItemTemplate} from './ActionListItem.stories'
-import {DividerTemplate} from './ActionListDivider.stories'
-import {ListTemplate} from './ActionList.stories'
+import {ListItemTemplate} from '../../private/ActionList/ActionListItem.stories'
+import {ListTemplate} from '../../private/ActionList/ActionList.stories.jsx'
+import {
+  TextOnly,
+  SizeMedium,
+  SizeMediumWithDescription,
+  SizeLarge,
+  SizeLargeWithDescription,
+  VisualLeading,
+  VisualTrailing,
+  VisualTrailingText,
+  VisualLeadingAndTrailing,
+  DescriptionBlock,
+  DescriptionBlockWithLeadingVisual,
+  DescriptionBlockWithTrailingVisual,
+  DescriptionBlockWithLeadingAndTrailingVisual,
+  DescriptionInline,
+  DescriptionInlineWithLeadingVisual,
+  DescriptionInlineWithTrailingVisual,
+  DescriptionInlineWithLeadingAndTrailingVisual,
+  NavActiveAnchor,
+  NavActivePage,
+  VariantDangerItem,
+  VariantDangerItemLeading,
+  VariantDangerItemTrailing,
+  VariantDangerItemLeadingTrailing,
+  TruncateText,
+  WrapText
+} from '../../private/ActionList/ActionListItemFeatures.stories.jsx'
 
 export default {
-  title: 'Patterns/ActionList/ActionListItem/Features',
+  title: 'Components/ActionList/ActionListItem/Features',
   parameters: {
     design: {
       type: 'figma',
@@ -35,325 +61,261 @@ const trailingVisual = `<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 16 
       ></path>
     </svg>`
 
-export const TextOnly = ListItemTemplate.bind({})
+export const TextOnly2 = TextOnly.bind({})
 TextOnly.args = {
-  text: 'Basic item label'
+  ...TextOnly.args
 }
 
-export const SizeMedium = ListItemTemplate.bind({})
-SizeMedium.args = {
-  ...ListItemTemplate.args,
-  text: 'Medium item',
-  size: 'ActionList-content--sizeMedium'
-}
+// export const SizeMedium = ListItemTemplate.bind({})
+// SizeMedium.args = {
+//   ...ListItemTemplate.args,
+//   text: 'Medium item',
+//   size: 'ActionList-content--sizeMedium'
+// }
 
-export const SizeMediumWithDescription = ListItemTemplate.bind({})
-SizeMediumWithDescription.args = {
-  ...ListItemTemplate.args,
-  text: 'Medium item',
-  description: 'Some descriptive text',
-  size: 'ActionList-content--sizeMedium'
-}
+// export const SizeMediumWithDescription = ListItemTemplate.bind({})
+// SizeMediumWithDescription.args = {
+//   ...ListItemTemplate.args,
+//   text: 'Medium item',
+//   description: 'Some descriptive text',
+//   size: 'ActionList-content--sizeMedium'
+// }
 
-export const SizeLarge = ListItemTemplate.bind({})
-SizeLarge.args = {
-  ...ListItemTemplate.args,
-  text: 'Large item',
-  size: 'ActionList-content--sizeLarge'
-}
+// export const SizeLarge = ListItemTemplate.bind({})
+// SizeLarge.args = {
+//   ...ListItemTemplate.args,
+//   text: 'Large item',
+//   size: 'ActionList-content--sizeLarge'
+// }
 
-export const SizeLargeWithDescription = ListItemTemplate.bind({})
-SizeLargeWithDescription.args = {
-  ...ListItemTemplate.args,
-  text: 'Large item',
-  description: 'Some descriptive text',
-  size: 'ActionList-content--sizeLarge'
-}
+// export const SizeLargeWithDescription = ListItemTemplate.bind({})
+// SizeLargeWithDescription.args = {
+//   ...ListItemTemplate.args,
+//   text: 'Large item',
+//   description: 'Some descriptive text',
+//   size: 'ActionList-content--sizeLarge'
+// }
 
-export const VisualLeading = ListItemTemplate.bind({})
-VisualLeading.storyName = '[Visuals] Leading'
-VisualLeading.args = {
-  ...ListItemTemplate.args,
-  text: 'Item with leading visual',
-  leadingVisual: leadingVisual
-}
+// export const VisualLeading = ListItemTemplate.bind({})
+// VisualLeading.storyName = '[Visuals] Leading'
+// VisualLeading.args = {
+//   ...ListItemTemplate.args,
+//   text: 'Item with leading visual',
+//   leadingVisual: leadingVisual
+// }
 
-export const VisualTrailing = ListItemTemplate.bind({})
-VisualTrailing.storyName = '[Visuals] Trailing'
-VisualTrailing.args = {
-  ...ListItemTemplate.args,
-  text: 'Item with trailing visual',
-  trailingVisual: trailingVisual
-}
+// export const VisualTrailing = ListItemTemplate.bind({})
+// VisualTrailing.storyName = '[Visuals] Trailing'
+// VisualTrailing.args = {
+//   ...ListItemTemplate.args,
+//   text: 'Item with trailing visual',
+//   trailingVisual: trailingVisual
+// }
 
-export const VisualTrailingText = ListItemTemplate.bind({})
-VisualTrailingText.storyName = '[Visuals] Trailing text'
-VisualTrailingText.args = {
-  ...ListItemTemplate.args,
-  text: 'Item with trailing text',
-  trailingVisual: `⌘N`
-}
+// export const VisualTrailingText = ListItemTemplate.bind({})
+// VisualTrailingText.storyName = '[Visuals] Trailing text'
+// VisualTrailingText.args = {
+//   ...ListItemTemplate.args,
+//   text: 'Item with trailing text',
+//   trailingVisual: `⌘N`
+// }
 
-export const VisualLeadingAndTrailing = ListItemTemplate.bind({})
-VisualLeadingAndTrailing.storyName = '[Visuals] Leading & trailing'
-VisualLeadingAndTrailing.args = {
-  ...ListItemTemplate.args,
-  text: 'Item with trailing visual',
-  trailingVisual: trailingVisual,
-  leadingVisual: leadingVisual
-}
+// export const VisualLeadingAndTrailing = ListItemTemplate.bind({})
+// VisualLeadingAndTrailing.storyName = '[Visuals] Leading & trailing'
+// VisualLeadingAndTrailing.args = {
+//   ...ListItemTemplate.args,
+//   text: 'Item with trailing visual',
+//   trailingVisual: trailingVisual,
+//   leadingVisual: leadingVisual
+// }
 
-export const DescriptionBlock = ListItemTemplate.bind({})
-DescriptionBlock.storyName = '[Description] block'
-DescriptionBlock.args = {
-  ...ListItemTemplate.args,
-  text: 'Item label',
-  description: 'This is a description',
-  descriptionVariant: 'ActionList-item-blockDescription'
-}
+// export const DescriptionBlock = ListItemTemplate.bind({})
+// DescriptionBlock.storyName = '[Description] block'
+// DescriptionBlock.args = {
+//   ...ListItemTemplate.args,
+//   text: 'Item label',
+//   description: 'This is a description',
+//   descriptionVariant: 'ActionList-item-blockDescription'
+// }
 
-export const DescriptionBlockWithLeadingVisual = ListItemTemplate.bind({})
-DescriptionBlockWithLeadingVisual.storyName = '[Description] block + leading visual'
-DescriptionBlockWithLeadingVisual.args = {
-  ...ListItemTemplate.args,
-  text: 'Item label',
-  description: 'This is a description',
-  descriptionVariant: 'ActionList-item-blockDescription',
-  leadingVisual: leadingVisual
-}
+// export const DescriptionBlockWithLeadingVisual = ListItemTemplate.bind({})
+// DescriptionBlockWithLeadingVisual.storyName = '[Description] block + leading visual'
+// DescriptionBlockWithLeadingVisual.args = {
+//   ...ListItemTemplate.args,
+//   text: 'Item label',
+//   description: 'This is a description',
+//   descriptionVariant: 'ActionList-item-blockDescription',
+//   leadingVisual: leadingVisual
+// }
 
-export const DescriptionBlockWithTrailingVisual = ListItemTemplate.bind({})
-DescriptionBlockWithTrailingVisual.storyName = '[Description] block + trailing visual'
-DescriptionBlockWithTrailingVisual.args = {
-  ...ListItemTemplate.args,
-  text: 'Item label',
-  description: 'This is a description',
-  descriptionVariant: 'ActionList-item-blockDescription',
-  trailingVisual: trailingVisual
-}
+// export const DescriptionBlockWithTrailingVisual = ListItemTemplate.bind({})
+// DescriptionBlockWithTrailingVisual.storyName = '[Description] block + trailing visual'
+// DescriptionBlockWithTrailingVisual.args = {
+//   ...ListItemTemplate.args,
+//   text: 'Item label',
+//   description: 'This is a description',
+//   descriptionVariant: 'ActionList-item-blockDescription',
+//   trailingVisual: trailingVisual
+// }
 
-export const DescriptionBlockWithLeadingAndTrailingVisual = ListItemTemplate.bind({})
-DescriptionBlockWithLeadingAndTrailingVisual.storyName = '[Description] block + leading/trailing visual'
-DescriptionBlockWithLeadingAndTrailingVisual.args = {
-  ...ListItemTemplate.args,
-  text: 'Item label',
-  description: 'This is a description',
-  descriptionVariant: 'ActionList-item-blockDescription',
-  trailingVisual: trailingVisual,
-  leadingVisual: leadingVisual
-}
+// export const DescriptionBlockWithLeadingAndTrailingVisual = ListItemTemplate.bind({})
+// DescriptionBlockWithLeadingAndTrailingVisual.storyName = '[Description] block + leading/trailing visual'
+// DescriptionBlockWithLeadingAndTrailingVisual.args = {
+//   ...ListItemTemplate.args,
+//   text: 'Item label',
+//   description: 'This is a description',
+//   descriptionVariant: 'ActionList-item-blockDescription',
+//   trailingVisual: trailingVisual,
+//   leadingVisual: leadingVisual
+// }
 
-export const DescriptionInline = ListItemTemplate.bind({})
-DescriptionInline.storyName = '[Description] inline'
-DescriptionInline.args = {
-  ...ListItemTemplate.args,
-  text: 'Item label',
-  description: 'This is a description',
-  descriptionVariant: 'ActionList-item-descriptionWrap--inline'
-}
+// export const DescriptionInline = ListItemTemplate.bind({})
+// DescriptionInline.storyName = '[Description] inline'
+// DescriptionInline.args = {
+//   ...ListItemTemplate.args,
+//   text: 'Item label',
+//   description: 'This is a description',
+//   descriptionVariant: 'ActionList-item-descriptionWrap--inline'
+// }
 
-export const DescriptionInlineWithLeadingVisual = ListItemTemplate.bind({})
-DescriptionInlineWithLeadingVisual.storyName = '[Description] inline + leading visual'
-DescriptionInlineWithLeadingVisual.args = {
-  ...ListItemTemplate.args,
-  text: 'Item label',
-  description: 'This is a description',
-  descriptionVariant: 'ActionList-item-descriptionWrap--inline',
-  leadingVisual: leadingVisual
-}
+// export const DescriptionInlineWithLeadingVisual = ListItemTemplate.bind({})
+// DescriptionInlineWithLeadingVisual.storyName = '[Description] inline + leading visual'
+// DescriptionInlineWithLeadingVisual.args = {
+//   ...ListItemTemplate.args,
+//   text: 'Item label',
+//   description: 'This is a description',
+//   descriptionVariant: 'ActionList-item-descriptionWrap--inline',
+//   leadingVisual: leadingVisual
+// }
 
-export const DescriptionInlineWithTrailingVisual = ListItemTemplate.bind({})
-DescriptionInlineWithTrailingVisual.storyName = '[Description] inline + trailing visual'
-DescriptionInlineWithTrailingVisual.args = {
-  ...ListItemTemplate.args,
-  text: 'Item label',
-  description: 'This is a description',
-  descriptionVariant: 'ActionList-item-descriptionWrap--inline',
-  trailingVisual: trailingVisual
-}
+// export const DescriptionInlineWithTrailingVisual = ListItemTemplate.bind({})
+// DescriptionInlineWithTrailingVisual.storyName = '[Description] inline + trailing visual'
+// DescriptionInlineWithTrailingVisual.args = {
+//   ...ListItemTemplate.args,
+//   text: 'Item label',
+//   description: 'This is a description',
+//   descriptionVariant: 'ActionList-item-descriptionWrap--inline',
+//   trailingVisual: trailingVisual
+// }
 
-export const DescriptionInlineWithLeadingAndTrailingVisual = ListItemTemplate.bind({})
-DescriptionInlineWithLeadingAndTrailingVisual.storyName = '[Description] inline + leading/trailing visual'
-DescriptionInlineWithLeadingAndTrailingVisual.args = {
-  ...ListItemTemplate.args,
-  text: 'Item label',
-  description: 'This is a description',
-  descriptionVariant: 'ActionList-item-descriptionWrap--inline',
-  trailingVisual: trailingVisual,
-  leadingVisual: leadingVisual
-}
+// export const DescriptionInlineWithLeadingAndTrailingVisual = ListItemTemplate.bind({})
+// DescriptionInlineWithLeadingAndTrailingVisual.storyName = '[Description] inline + leading/trailing visual'
+// DescriptionInlineWithLeadingAndTrailingVisual.args = {
+//   ...ListItemTemplate.args,
+//   text: 'Item label',
+//   description: 'This is a description',
+//   descriptionVariant: 'ActionList-item-descriptionWrap--inline',
+//   trailingVisual: trailingVisual,
+//   leadingVisual: leadingVisual
+// }
 
-export const NavActiveAnchor = ListItemTemplate.bind({})
-NavActiveAnchor.storyName = '[Nav] Active anchor'
-NavActiveAnchor.args = {
-  ...ListItemTemplate.args,
-  text: 'Im an anchor link',
-  href: '#someid',
-  ariaCurrent: 'location',
-  activeNavItem: true
-}
+// export const NavActiveAnchor = ListItemTemplate.bind({})
+// NavActiveAnchor.storyName = '[Nav] Active anchor'
+// NavActiveAnchor.args = {
+//   ...ListItemTemplate.args,
+//   text: 'Im an anchor link',
+//   href: '#someid',
+//   ariaCurrent: 'location',
+//   activeNavItem: true
+// }
 
-export const NavActivePage = ListItemTemplate.bind({})
-NavActivePage.storyName = '[Nav] Active page'
-NavActivePage.args = {
-  ...ListItemTemplate.args,
-  text: 'Im a page level link',
-  href: '/',
-  ariaCurrent: 'page',
-  activeNavItem: true
-}
+// export const NavActivePage = ListItemTemplate.bind({})
+// NavActivePage.storyName = '[Nav] Active page'
+// NavActivePage.args = {
+//   ...ListItemTemplate.args,
+//   text: 'Im a page level link',
+//   href: '/',
+//   ariaCurrent: 'page',
+//   activeNavItem: true
+// }
 
-export const VariantDangerItem = ListItemTemplate.bind({})
-VariantDangerItem.storyName = '[Variant] Danger'
-VariantDangerItem.args = {
-  ...ListItemTemplate.args,
-  text: 'Danger danger',
-  variant: 'ActionList-item--danger'
-}
+// export const VariantDangerItem = ListItemTemplate.bind({})
+// VariantDangerItem.storyName = '[Variant] Danger'
+// VariantDangerItem.args = {
+//   ...ListItemTemplate.args,
+//   text: 'Danger danger',
+//   variant: 'ActionList-item--danger'
+// }
 
-export const VariantDangerItemLeading = ListItemTemplate.bind({})
-VariantDangerItemLeading.storyName = '[Variant] Danger + leading visual'
-VariantDangerItemLeading.args = {
-  ...ListItemTemplate.args,
-  text: 'Danger danger',
-  variant: 'ActionList-item--danger',
-  leadingVisual: leadingVisual
-}
+// export const VariantDangerItemLeading = ListItemTemplate.bind({})
+// VariantDangerItemLeading.storyName = '[Variant] Danger + leading visual'
+// VariantDangerItemLeading.args = {
+//   ...ListItemTemplate.args,
+//   text: 'Danger danger',
+//   variant: 'ActionList-item--danger',
+//   leadingVisual: leadingVisual
+// }
 
-export const VariantDangerItemTrailing = ListItemTemplate.bind({})
-VariantDangerItemTrailing.storyName = '[Variant] Danger + trailing visual'
-VariantDangerItemTrailing.args = {
-  ...ListItemTemplate.args,
-  text: 'Danger danger',
-  variant: 'ActionList-item--danger',
-  trailingVisual: trailingVisual
-}
+// export const VariantDangerItemTrailing = ListItemTemplate.bind({})
+// VariantDangerItemTrailing.storyName = '[Variant] Danger + trailing visual'
+// VariantDangerItemTrailing.args = {
+//   ...ListItemTemplate.args,
+//   text: 'Danger danger',
+//   variant: 'ActionList-item--danger',
+//   trailingVisual: trailingVisual
+// }
 
-export const VariantDangerItemLeadingTrailing = ListItemTemplate.bind({})
-VariantDangerItemLeadingTrailing.storyName = '[Variant] Danger + leading/trailing visual'
-VariantDangerItemLeadingTrailing.args = {
-  ...ListItemTemplate.args,
-  text: 'Danger danger',
-  variant: 'ActionList-item--danger',
-  trailingVisual: trailingVisual,
-  leadingVisual: leadingVisual
-}
+// export const VariantDangerItemLeadingTrailing = ListItemTemplate.bind({})
+// VariantDangerItemLeadingTrailing.storyName = '[Variant] Danger + leading/trailing visual'
+// VariantDangerItemLeadingTrailing.args = {
+//   ...ListItemTemplate.args,
+//   text: 'Danger danger',
+//   variant: 'ActionList-item--danger',
+//   trailingVisual: trailingVisual,
+//   leadingVisual: leadingVisual
+// }
 
-export const VariantDisabledItem = ListItemTemplate.bind({})
-VariantDisabledItem.storyName = '[Variant] Disabled'
-VariantDisabledItem.args = {
-  ...ListItemTemplate.args,
-  text: 'Disabled',
-  ariaDisabled: true
-}
+// export const VariantDisabledItem = ListItemTemplate.bind({})
+// VariantDisabledItem.storyName = '[Variant] Disabled'
+// VariantDisabledItem.args = {
+//   ...ListItemTemplate.args,
+//   text: 'Disabled',
+//   ariaDisabled: true
+// }
 
-export const VariantDisabledItemLeading = ListItemTemplate.bind({})
-VariantDisabledItemLeading.storyName = '[Variant] Disabled + leading visual'
-VariantDisabledItemLeading.args = {
-  ...ListItemTemplate.args,
-  text: 'Disabled',
-  ariaDisabled: true,
-  leadingVisual: leadingVisual
-}
+// export const VariantDisabledItemLeading = ListItemTemplate.bind({})
+// VariantDisabledItemLeading.storyName = '[Variant] Disabled + leading visual'
+// VariantDisabledItemLeading.args = {
+//   ...ListItemTemplate.args,
+//   text: 'Disabled',
+//   ariaDisabled: true,
+//   leadingVisual: leadingVisual
+// }
 
-export const VariantDisabledItemTrailing = ListItemTemplate.bind({})
-VariantDisabledItemTrailing.storyName = '[Variant] Disabled + trailing visual'
-VariantDisabledItemTrailing.args = {
-  ...ListItemTemplate.args,
-  text: 'Disabled',
-  ariaDisabled: true,
-  trailingVisual: trailingVisual
-}
+// export const VariantDisabledItemTrailing = ListItemTemplate.bind({})
+// VariantDisabledItemTrailing.storyName = '[Variant] Disabled + trailing visual'
+// VariantDisabledItemTrailing.args = {
+//   ...ListItemTemplate.args,
+//   text: 'Disabled',
+//   ariaDisabled: true,
+//   trailingVisual: trailingVisual
+// }
 
-export const VariantDisabledItemLeadingTrailing = ListItemTemplate.bind({})
-VariantDisabledItemLeadingTrailing.storyName = '[Variant] Disabled + leading/trailing visual'
-VariantDisabledItemLeadingTrailing.args = {
-  ...ListItemTemplate.args,
-  text: 'Disabled',
-  ariaDisabled: true,
-  trailingVisual: trailingVisual,
-  leadingVisual: leadingVisual
-}
+// export const VariantDisabledItemLeadingTrailing = ListItemTemplate.bind({})
+// VariantDisabledItemLeadingTrailing.storyName = '[Variant] Disabled + leading/trailing visual'
+// VariantDisabledItemLeadingTrailing.args = {
+//   ...ListItemTemplate.args,
+//   text: 'Disabled',
+//   ariaDisabled: true,
+//   trailingVisual: trailingVisual,
+//   leadingVisual: leadingVisual
+// }
 
-export const ActionSingleSelectItem = ListItemTemplate.bind({})
-ActionSingleSelectItem.storyName = '[Actions] Single select'
-ActionSingleSelectItem.args = {
-  text: 'Single select item',
-  singleSelect: true
-}
+// export const TruncateText = ListItemTemplate.bind({})
+// TruncateText.storyName = 'Truncate text'
+// TruncateText.args = {
+//   ...ListItemTemplate.args,
+//   truncateItem: true,
+//   text: 'This is a very long string of text that will wrap to another line on smaller screens',
+//   trailingVisual: trailingVisual,
+//   leadingVisual: leadingVisual
+// }
 
-export const ActionSingleSelectItemWithLeadingVisual = ListItemTemplate.bind({})
-ActionSingleSelectItemWithLeadingVisual.storyName = '[Actions] Single select + leading visual'
-ActionSingleSelectItemWithLeadingVisual.args = {
-  text: 'Single select item',
-  singleSelect: true,
-  leadingVisual: leadingVisual
-}
-
-export const ActionSingleSelectItemWithTrailingVisual = ListItemTemplate.bind({})
-ActionSingleSelectItemWithTrailingVisual.storyName = '[Actions] Single select + trailing visual'
-ActionSingleSelectItemWithTrailingVisual.args = {
-  text: 'Single select item',
-  singleSelect: true,
-  trailingVisual: trailingVisual
-}
-
-export const ActionSingleSelectItemWithLeadingAndTrailingVisual = ListItemTemplate.bind({})
-ActionSingleSelectItemWithLeadingAndTrailingVisual.storyName = '[Actions] Single select + leading/trailing visual'
-ActionSingleSelectItemWithLeadingAndTrailingVisual.args = {
-  text: 'Single select item',
-  singleSelect: true,
-  trailingVisual: trailingVisual,
-  leadingVisual: leadingVisual
-}
-
-export const ActionMultiSelectItem = ListItemTemplate.bind({})
-ActionMultiSelectItem.storyName = '[Actions] Multi select'
-ActionMultiSelectItem.args = {
-  text: 'Multi select item',
-  multiSelect: true
-}
-
-export const ActionMultiSelectItemWithLeadingVisual = ListItemTemplate.bind({})
-ActionMultiSelectItemWithLeadingVisual.storyName = '[Actions] Multi select + leading visual'
-ActionMultiSelectItemWithLeadingVisual.args = {
-  text: 'Multi select item',
-  multiSelect: true,
-  leadingVisual: leadingVisual
-}
-
-export const ActionMultiSelectItemWithTrailingVisual = ListItemTemplate.bind({})
-ActionMultiSelectItemWithTrailingVisual.storyName = '[Actions] Multi select + trailing visual'
-ActionMultiSelectItemWithTrailingVisual.args = {
-  text: 'Multi select item',
-  multiSelect: true,
-  trailingVisual: trailingVisual
-}
-
-export const ActionMultiSelectItemWithLeadingAndTrailingVisual = ListItemTemplate.bind({})
-ActionMultiSelectItemWithLeadingAndTrailingVisual.storyName = '[Actions] Multi select + leading/trailing visual'
-ActionMultiSelectItemWithLeadingAndTrailingVisual.args = {
-  text: 'Multi select item',
-  multiSelect: true,
-  trailingVisual: trailingVisual,
-  leadingVisual: leadingVisual
-}
-
-export const TruncateText = ListItemTemplate.bind({})
-TruncateText.storyName = 'Truncate text'
-TruncateText.args = {
-  ...ListItemTemplate.args,
-  truncateItem: true,
-  text: 'This is a very long string of text that will wrap to another line on smaller screens',
-  trailingVisual: trailingVisual,
-  leadingVisual: leadingVisual
-}
-
-export const WrapText = ListItemTemplate.bind({})
-WrapText.storyName = 'Wrap text'
-WrapText.args = {
-  ...ListItemTemplate.args,
-  text: 'This is a very long string of text that will wrap to another line on smaller screens',
-  trailingVisual: trailingVisual,
-  leadingVisual: leadingVisual
-}
+// export const WrapText = ListItemTemplate.bind({})
+// WrapText.storyName = 'Wrap text'
+// WrapText.args = {
+//   ...ListItemTemplate.args,
+//   text: 'This is a very long string of text that will wrap to another line on smaller screens',
+//   trailingVisual: trailingVisual,
+//   leadingVisual: leadingVisual
+// }
