+++ docs/src/stories/private/ActionList/API.stories.mdx
@@ -1,8 +1,10 @@
 import {Meta, Story, Canvas} from '@storybook/addon-docs'
 
-<Meta title="Patterns/ActionList/Accessibility" />
+<Meta title="Private/ActionList/API" />
 
-### Action List
+### Action list
+
+`ActionList` CSS is primarily used as the building blocks for compositional components, such as NavigationList. Below is an overview of how the markup may look for specific kinds of implementions and contexts.
 
 An action list is a vertical list of interactive <strong>actions</strong> or <strong>options</strong>.
 
@@ -13,13 +15,13 @@ An action list is a vertical list of interactive <strong>actions</strong> or <st
 
 ### Semantic markup
 
-The markup for Action List changes depending on the intended use case.
+The markup for Action list changes depending on the intended use case.
 
 In all cases, the basic structure is as follows:
 
 ```html
 <ul>
-  <li><span>Content with action</span></li>
+  <li><button>Content with action</button></li>
   <li><a href="/">Content as link</a></li>
   <li>
     <ul>
@@ -29,13 +31,11 @@ In all cases, the basic structure is as follows:
 </ul>
 ```
 
-Pay close attention to `role` attributes throughout the documentation, and find the use case that best suits your needs.
-
-### Dropdown Menu
+### Menu
 
 [Menu spec](https://www.w3.org/TR/wai-aria-practices-1.2/examples/menu-button/menu-button-links.html)
 
-The `menu` role should be reserved for dropdown menu scenarios
+The `menu` role should be reserved for menu scenarios
 
 | Element | Role                            |
 | :------ | :------------------------------ |
@@ -50,9 +50,14 @@ The `menu` role should be reserved for dropdown menu scenarios
 
 Wrap `ActionList` in a `nav` element for site or page navigation, like a sidebar
 
-No roles needed
+| Element | Role          |
+| :------ | :------------ |
+| `ul`    | `role="list"` |
+| `li`    | no role       |
+| `li`    | no role       |
+| `a`     | no role       |
 
-### Multi select dropdown menu
+### Multi select menu
 
 [Menu spec](https://www.w3.org/TR/wai-aria-practices-1.2/examples/menubar/menubar-2/menubar-2.html)
 
@@ -62,7 +67,7 @@ No roles needed
 | `li`    | `role="menuitemcheckbox"`   |
 | `li`    | `aria-checked="true/false"` |
 
-### Single select dropdown menu
+### Single select menu
 
 [Menu spec](https://www.w3.org/TR/wai-aria-practices-1.2/examples/menubar/menubar-2/menubar-2.html)
 
@@ -97,12 +102,10 @@ Note: JS is required for to provide [keyboard handling](https://www.w3.org/TR/wa
 
 ### Dividers
 
-| Element | Role               |
-| :------ | :----------------- |
-| `li`    | `role="separator"` |
+| Element | Role                                    |
+| :------ | :-------------------------------------- |
+| `li`    | `role="separator"` `aria-hidden="true"` |
 
-| Element | Role                                        |
-| :------ | :------------------------------------------ |
-| `li`    | `role="presentation"`                       |
-| `li`    | id for nested group `id="group-id"`         |
-| `li`    | if it has id for group `aria-hidden="true"` |
+| Element | Role |
+| :------ | :--- |
+| `h3`    | id   |
