+++ src/actionlist/action-list.scss
@@ -8,6 +8,11 @@
   padding: 0;
 }
 
+// nested list within main ActionList
+.ActionList--subGroup {
+  padding: 0;
+}
+
 // dividers
 
 .ActionList--divided {
